web/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
djangorestframework==3.13.1
oauth2-provider==0.0
pytz==2021.3
sqlparse==0.4.2
